FIT-31: Make accordion component more abstract
